refactor: integrate event generator into the calendar action menu
